requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
email-validator
pydantic
regex
PyJWT